Remove // comments on all lines
In some cases the diff will put a "+" line after a line-ending comment line, e.g.:
```
  // Foo
+private int field;
``` 
In this case, the field can accidentally get commented out by the removal of the line ending between the two lines. We already have a sed command that handles this case in lines beginning with "+", where we replace `//` with `/*` and add `/**/` to the end of the line, so we can generalize this to all lines to avoid this problem.